Implement and test the greeting query
This commit adds a `GreetingController` and a `@QueryMapping` annotated
handler for the "greeting" query. This is also refining the GraphQL
schema to improve argument binding and default values.

We can test this new mapping with the following query:

```
{
  greeting(name: "Brian")
}
```

This is also showing how we can test this implementation with a mock web
layer and a GraphQL test client.