Link ArtifactRepository to Releases
Because Project Releases are hosted in artifact repositories, we could
improve our object graph and link releases to their respective artifact
repository.

This commit expresses the relationship between a Release and its
Repository using a `@SchemaMapping` annotation.
The method signature gives enough information to the engine for
materializing that relationship.

We can now request the repository information in our previous query:

```
{
  project(slug: "spring-graphql") {
    name
    status
    releases {
      version
      status
      repository {
        url
      }
    }
  }
}
```