Expose Contributor and GitHub Profile information
This commit uses the newly introduced GitHub client to fetch information
about contributors and their GitHub profiles.

For that, we need to implement the relationship between `Project` and
`Contributor` using `@SchemaMapping`. We could apply the same reasoning
for the `Contributor` to `Profile` relationship, but in this case we
could optimize for batch loading. While this is not the case yet, as
we're using GitHub's REST API, this setup could be still useful if the
same contributor is fetched multiple times in the same request, or if we
use GitHub's GraphQL API in the future.

We can test this new feature with the following query:

```
{
  project(slug: "spring-graphql") {
    name
    status
    contributors {
      gitHubId
      profile {
        name
      }
    }
  }
}
```